Rename debug.ts helper file to logging.ts
